% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bagui.R
\name{bagui}
\alias{bagui}
\title{bagui}
\usage{
bagui()
}
\description{
Launches GUI to make using this package really, really easy
}
\details{
None at this time
}
\author{
Jung-han Wang and Robert Norberg
}

