% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data Documentation.R
\name{Before}
\alias{Before}
\title{Traffic accidents at an intersection prior to the installation of a traffic signal}
\format{A data frame with 228 rows and 4 variables}
\description{
A dataset containing the number of traffic accidents at an intersection
The variables are as follows:
}
\details{
\itemize{
  \item kabco. The combined number of accidents of several classes (k = fatality, a = incapacitating injury, b = non-incapacitating injury, c = possible injury, o = property damage only)
  \item Max_AADT. Maximum Annual Average Daily Traffic
  \item Min_AADT. Minimum Annual Average Daily Traffic
  \item year. number of years of data included in the data set
}
}

