% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPredVals.R
\name{calcPredVals}
\alias{calcPredVals}
\alias{fitReferenceModel}
\title{calcPredVals}
\usage{
calcPredVals(refMod, trtData, BAcol = "BeforeAfter")
}
\arguments{
\item{refMod}{A model object created by fitReferenceModel}

\item{trtData}{A data.frame of treatment crash data.
This data.frame should have column names that exactly match those in mod$model.
It should have an additional column indicating before/after.}

\item{BAcol}{The name of the column indicating before/after. This column should have no missing data and each value should be one of "Before" or "After".}
}
\value{
Returns the treatment data frame with an additional column named "Predicted".
}
\description{
xxx
}
\details{
None at this time
}
\examples{
data(Reference)
ref_mod <- fitReferenceModel(refData = Reference,
 depVar = "kabco_0312",
 indepVars = c("Max_AADT", "Min_AADT"),
 offset="year")
}
\author{
Jung-han Wang and Robert Norberg
}

