% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comparison.R
\name{CompGroup}
\alias{CompGroup}
\title{CompGroup}
\usage{
CompGroup(compBefore, compAfter, before, after, depVar, alpha = 0.95)
}
\arguments{
\item{compBefore}{Comparison data in before period}

\item{compAfter}{Comparison data in after period}

\item{before}{Treatment data, before some change was made}

\item{after}{Treatment data, after some change was made}

\item{depVar}{The dependent variable (the number of crashes - should always be of class integer or numeric).}

\item{alpha}{Level of confidence}
}
\value{
Returns a list object containing the CMF, its variance, standard error, and 1-alpha/2 CI
}
\description{
Calculates Crash Modification Factor (CMF) for a before/after traffic study using Comparison Method
}
\details{
None at this time
}
\examples{
data(compBefore)
data(compAfter)
data(Before)
data(After)
CompGroup(compBefore = compBefore, compAfter = compAfter, before = Before, after = After,
 depVar = "kabco")
}
\author{
Jung-han Wang and Robert Norberg
}

