% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/empBayes.R
\name{empBayes}
\alias{empBayes}
\alias{fitReferenceModel}
\title{calcPredVals}
\usage{
empBayes(trtData, refModel, BAcol = "BeforeAfter", alpha = 0.95)
}
\arguments{
\item{trtData}{Treatment data returned from calling calcPredVals()}

\item{refModel}{Reference model returned from calling fitReferencModel()}

\item{BAcol}{The name of the column indicating before/after. This column should have no missing data and each value should be one of "Before" or "After".}

\item{alpha}{Level of confidence}
}
\value{
Returns the treatment data frame with an additional column named "Predicted".
}
\description{
xxx
}
\details{
None at this time
}
\examples{
ref_mod <- fitReferenceModel(refData = Reference,
 depVar = "kabco_0312",
 indepVars = c("Max_AADT", "Min_AADT"),
 offset="year")
}
\author{
Jung-han Wang and Robert Norberg
}

