% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitReferenceModel.R
\name{fitReferenceModel}
\alias{fitReferenceModel}
\title{fitReferenceModel}
\usage{
fitReferenceModel(refData, depVar, indepVars, offset = NULL)
}
\arguments{
\item{refData}{A data.frame of reference crash data}

\item{depVar}{The dependent variable (the number of crashes - should always be of class integer or numeric).}

\item{indepVars}{Variables used to model the outcome variable depVar}

\item{offset}{An offset variable (eg years)}
}
\value{
Returns a negative binomial model object
}
\description{
Fits a negative binomial model to the reference data
}
\details{
None at this time
}
\examples{
fitReferenceModel(refData = Reference,
 depVar = "kabco_0312",
 indepVars = c("Max_AADT", "Min_AADT"),
 offset="year")
}
\author{
Jung-han Wang and Robert Norberg
}

