% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naive.r
\name{naive}
\alias{naive}
\title{naive}
\usage{
naive(before, after, depVar, db = 1, da = 1, alpha = 0.95)
}
\arguments{
\item{before}{Treatment data, before some change was made}

\item{after}{Treatment data, after some change was made}

\item{depVar}{The dependent variable (the number of crashes - should always be of class integer or numeric).}

\item{db}{duration of before period (typically years)}

\item{da}{duration of after period (typically years)}

\item{alpha}{Level of confidence}
}
\value{
Returns a list object containing the CMF, its variance, standard error, and 1-alpha/2 CI
}
\description{
Calculates Crash Modification Factor (CMF) for a before/after traffic study using Naive method
}
\details{
None at this time
}
\examples{
data(Before)
data(After)
naive(before = Before, after = After, depVar = "kabco", db = 1, da = 3, alpha = 0.9)
}
\author{
Jung-han Wang and Robert Norberg
}

