% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Misc_Documentation.R
\name{Reference}
\alias{Reference}
\title{Traffic accidents at a reference site from 2003 to 2012}
\format{A data frame with 318 rows and 4 variables}
\description{
A dataset containing the number of traffic accidents at an intersection
The variables are as follows:
}
\details{
\itemize{
  \item kabco. The combined number of accidents of several classes (k = fatality, a = incapacitating injury, b = non-incapacitating injury, c = possible injury, o = property damage only)
  \item Max_AADT. Maximum Annual Average Daily Traffic
  \item Min_AADT. Minimum Annual Average Daily Traffic
  \item year. number of years of data included in the data set
}
}

